requirements.txt
requests==2.*
mysql-connector-python==8.*
npyscreen==4.*
